azure_tenant_refresh/tests/requirements.txt
pytest==3.6.3
azure-mgmt-authorization==0.30.0
azure-mgmt-monitor==0.5.2
azure-mgmt-resource==2.0.0